sharp-primer-designer/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]
primer3-py>=2.0.0
biopython>=1.80
python-multipart
python-dotenv
pydantic>=2.0
pytest
httpx
